src/requirements.txt
google-api-python-client>=1.12.8
google-cloud-bigquery>=2.9.0
google-cloud-storage>=1.36.0
google-resumable-media>=1.2.0
google-generativeai
google-cloud-documentai
oauth2client>=4.1.3
gspread>=3.6.0
gcsfs>=2022.2.0
google-cloud-aiplatform>=1.68.0
hugchat
markdown

IPython
python-docx
python-dotenv
PyPDF2
crcmod
langchain==0.2.16
langchain-community==0.2.17
streamlit-utils
streamlit-modal
streamlit-authenticator
watchdog
openpyxl
tabulate


